requirements.txt
django==4.2.4
dj-database-url==2.0.0
psycopg2-binary==2.9.7
whitenoise[brotli]==6.5.0
gunicorn==21.2.0
pillow==10.0.0
six==1.15.0
django-haystack==3.2.1
whoosh==2.7.4
